download dropdown to 100% thrusters
