+++ dash/webapp/pages/toggles.js
@@ -26,16 +26,7 @@ function setCookie(name, value, domain = 'wellcomecollection.org') {
     ''}; Path=/; Domain=${domain}; ${expiration}`;
 }
 
-const abTests = [
-  {
-    id: 'useBetaDownloadComponent',
-    title: 'Use the beta download component',
-    description:
-      'Use the beta version of the download component, ' +
-      'which will be more compatible with multiple download options in the future.' +
-      "We first want to check it doesn't negatively effect current download behaviour.",
-  },
-];
+const abTests = [];
 
 const featureToggles = [
   {
